Add fragments to home page
