+++ lib/main.dart
@@ -1,8 +1,6 @@
 // 🌎 Project imports:
 import 'package:enzitech_app/src/app_config.dart';
 
-// 🌎 Project imports:
-
 Future<void> main() async {
   initApp();
 }
