+++ lib/src/features/home/fragments/account/account_page.dart
@@ -1,0 +1,49 @@
+// 🐦 Flutter imports:
+import 'package:flutter/material.dart';
+
+// 📦 Package imports:
+import 'package:provider/provider.dart';
+
+// 🌎 Project imports:
+import 'package:enzitech_app/src/features/home/home_controller.dart';
+
+class AccountPage extends StatefulWidget {
+  const AccountPage({Key? key}) : super(key: key);
+
+  @override
+  State<AccountPage> createState() => _AccountPageState();
+}
+
+class _AccountPageState extends State<AccountPage> {
+  late final HomeController controller;
+
+  @override
+  void initState() {
+    super.initState();
+    controller = context.read<HomeController>();
+    if (mounted) {
+      controller.addListener(() {
+        if (controller.state == HomeState.error) {
+          ScaffoldMessenger.of(context).showSnackBar(
+            SnackBar(
+              content: Text(controller.failure!.message),
+            ),
+          );
+        }
+      });
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    var widthMQ = MediaQuery.of(context).size.width;
+    var heightMQ = MediaQuery.of(context).size.height;
+    final controller = context.watch<HomeController>();
+
+    return Container(
+      color: Colors.blue,
+      height: heightMQ,
+      width: widthMQ,
+    );
+  }
+}
