+++ lib/src/features/home/home_page.dart
@@ -5,15 +5,14 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:phosphor_flutter/phosphor_flutter.dart';
 import 'package:provider/provider.dart';
-import 'package:toggle_switch/toggle_switch.dart';
 
 // 🌎 Project imports:
 import 'package:enzitech_app/src/features/home/components/experiment_card.dart';
+import 'package:enzitech_app/src/features/home/fragments/account/account_page.dart';
+import 'package:enzitech_app/src/features/home/fragments/experiments/experiments_page.dart';
+import 'package:enzitech_app/src/features/home/fragments/treatments/treatments_page.dart';
 import 'package:enzitech_app/src/features/home/home_controller.dart';
-import 'package:enzitech_app/src/shared/validator/validator.dart';
-import '../../shared/themes/app_colors.dart';
-import '../../shared/themes/app_text_styles.dart';
-import '../../shared/widgets/ezt_textfield.dart';
+import 'package:enzitech_app/src/shared/themes/app_complete_theme.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage({Key? key}) : super(key: key);
@@ -25,9 +24,11 @@ class HomePage extends StatefulWidget {
 class _HomePageState extends State<HomePage> {
   late final HomeController controller;
 
-  final _searchTermController = TextEditingController(text: '');
-  List<bool> isSelected = [true, false];
-  int _selectedIndex = 0;
+  final List<Widget> _fragments = const [
+    ExperimentsPage(),
+    TreatmentsPage(),
+    AccountPage(),
+  ];
 
   @override
   void initState() {
@@ -46,97 +47,8 @@ class _HomePageState extends State<HomePage> {
     }
   }
 
-  void _onItemTapped(int index) {
-    setState(() {
-      _selectedIndex = index;
-    });
-  }
-
-  Widget get _searchTermInput {
-    final validations = <ValidateRule>[
-      ValidateRule(
-        ValidateTypes.name,
-      ),
-    ];
-
-    final fieldValidator = FieldValidator(validations, context);
-
-    return EZTTextField(
-      controller: _searchTermController,
-      lineHeight: 1.8,
-      eztTextFieldType: EZTTextFieldType.underline,
-      hintText: "Pesquisar experimento",
-      usePrimaryColorOnFocusedBorder: true,
-      keyboardType: TextInputType.name,
-      suffixIcon: const Icon(
-        PhosphorIcons.magnifyingGlass,
-        color: AppColors.primary,
-        size: 35,
-      ),
-      fieldValidator: fieldValidator,
-    );
-  }
-
-  Widget _buildExperimentsList(double height) {
-    if (controller.state == HomeState.loading) {
-      return Column(
-        children: [
-          SizedBox(
-            height: height / 3,
-          ),
-          const CircularProgressIndicator(),
-          SizedBox(
-            height: height / 3,
-          ),
-        ],
-      );
-    }
-
-    return Visibility(
-      visible: controller.mockedList.isNotEmpty,
-      replacement: Column(
-        children: [
-          SizedBox(height: height / 4),
-          Column(
-            children: const [
-              Icon(PhosphorIcons.folderDotted, size: 64),
-              Center(
-                child: Padding(
-                  padding: EdgeInsets.symmetric(horizontal: 35.0),
-                  child: Text(
-                    "Não existem experimentos!",
-                    textAlign: TextAlign.center,
-                    style: TextStyle(
-                      fontWeight: FontWeight.bold,
-                      fontSize: 24,
-                      color: AppColors.greyDark,
-                    ),
-                  ),
-                ),
-              ),
-            ],
-          ),
-          SizedBox(height: height / 4)
-        ],
-      ),
-      child: ListView.builder(
-        shrinkWrap: true,
-        physics: const BouncingScrollPhysics(),
-        itemCount: controller.mockedList.length,
-        itemBuilder: (context, index) {
-          return Padding(
-            padding: const EdgeInsets.only(bottom: 8.0),
-            child: controller.mockedList[index],
-          );
-        },
-      ),
-    );
-  }
-
   @override
   Widget build(BuildContext context) {
-    var widthMQ = MediaQuery.of(context).size.width;
-    var heightMQ = MediaQuery.of(context).size.height;
     final controller = context.watch<HomeController>();
 
     return Scaffold(
@@ -147,75 +59,39 @@ class _HomePageState extends State<HomePage> {
           alignment: Alignment.center,
         ),
       ),
-      body: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 8),
-        child: Column(
-          children: [
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 8.0),
-              child: _searchTermInput,
-            ),
-            const SizedBox(
-              height: 16,
-            ),
-            ToggleSwitch(
-              minWidth: widthMQ,
-              totalSwitches: 2,
-              labels: const ['Em andamento', 'Concluído'],
-              activeFgColor: AppColors.white,
-              inactiveFgColor: AppColors.primary,
-              activeBgColor: const [AppColors.primary],
-              inactiveBgColor: AppColors.white,
-              borderColor: const [AppColors.primary],
-              borderWidth: 1.5,
-              onToggle: (index) {
-                print(index);
-                // call controller to update search when this changes
+      body: ChangeNotifierProvider(
+        create: (BuildContext context) {},
+        child: _fragments[controller.fragmentIndex],
+      ),
+      floatingActionButton: controller.fragmentIndex == 0
+          ? FloatingActionButton.extended(
+              onPressed: () {
+                //TODO: REMOVER ESTE TESTE
+                if (controller.mockedList.isEmpty) {
+                  controller.setMockedList([
+                    ExperimentCard(
+                      name: 'Experimento 1',
+                      modifiedAt: DateTime.now(),
+                      description:
+                          'Esta é uma descrição opcional muito grande de experimento, bem detalhado, com muitas linhas, onde será permitido no máximo quatro linhas...',
+                      progress: .55,
+                    ),
+                  ]);
+                } else {
+                  controller.setMockedList([]);
+                }
               },
-            ),
-            const SizedBox(
-              height: 16,
-            ),
-            Expanded(
-              child: ScrollConfiguration(
-                behavior: const ScrollBehavior(),
-                child: GlowingOverscrollIndicator(
-                  axisDirection: AxisDirection.down,
-                  color: AppColors.primary,
-                  child: _buildExperimentsList(heightMQ),
-                ),
+              label: Text(
+                "Cadastrar\nexperimento",
+                style: TextStyles.buttonBackground,
               ),
-            ),
-          ],
-        ),
-      ),
-      floatingActionButton: FloatingActionButton.extended(
-        onPressed: () {
-          //TODO: REMOVER ESTE TESTE
-          if (controller.mockedList.isEmpty) {
-            controller.setMockedList([
-              ExperimentCard(
-                name: 'Experimento 1',
-                modifiedAt: DateTime.now(),
-                description:
-                    'Esta é uma descrição opcional muito grande de experimento, bem detalhado, com muitas linhas, onde será permitido no máximo quatro linhas...',
-                progress: .55,
+              icon: const Icon(
+                PhosphorIcons.pencilLine,
+                color: AppColors.white,
+                size: 30,
               ),
-            ]);
-          } else {
-            controller.setMockedList([]);
-          }
-        },
-        label: Text(
-          "Cadastrar\nexperimento",
-          style: TextStyles.buttonBackground,
-        ),
-        icon: const Icon(
-          PhosphorIcons.pencilLine,
-          color: AppColors.white,
-          size: 30,
-        ),
-      ),
+            )
+          : null,
       bottomNavigationBar: BottomNavigationBar(
         items: const [
           BottomNavigationBarItem(
@@ -228,13 +104,13 @@ class _HomePageState extends State<HomePage> {
           ),
           BottomNavigationBarItem(
             icon: Icon(PhosphorIcons.userCircleGear),
-            label: 'Perfil',
+            label: 'Conta',
           ),
         ],
-        currentIndex: _selectedIndex,
+        currentIndex: controller.fragmentIndex,
         selectedItemColor: AppColors.white,
         backgroundColor: AppColors.primary,
-        onTap: _onItemTapped,
+        onTap: (index) => controller.setFragmentIndex(index),
       ),
     );
   }
